feat: window class and console
- Create window class and editorWindow class.
- Start with docking.
- Create Console window.
- Capture DX12 messages.

+++ Source/ChironEngine/Source/DataModels/DX12/Resource/ConstantBuffer.cpp
@@ -27,7 +27,7 @@ void ConstantBuffer::CreateView()
 
     D3D12_CONSTANT_BUFFER_VIEW_DESC cbvDesc = {};
     cbvDesc.BufferLocation = _resource->GetGPUVirtualAddress();
-    cbvDesc.SizeInBytes = Chiron::Utils::AlignUp(_sizeInBytes, D3D12_CONSTANT_BUFFER_DATA_PLACEMENT_ALIGNMENT);
+    cbvDesc.SizeInBytes = static_cast<UINT>(Chiron::Utils::AlignUp(_sizeInBytes, D3D12_CONSTANT_BUFFER_DATA_PLACEMENT_ALIGNMENT));
 
     // Crear la CBV
     device->CreateConstantBufferView(&cbvDesc, _constantBufferView.GetCPUDescriptorHandle());
